renamed mlflow package to lab-mlflow-manager
