+++ components/lab-mlflow-manager/docker/nginx/nginx.conf
