+++ components/lab-mlflow-manager/webapp/public/manifest.json
